requirements.txt
faster-whisper>=0.10.0
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0
pathlib2>=2.3.0; python_version < "3.4"
click>=8.0.0
tqdm>=4.64.0
PyYAML>=6.0
